+++ modules/core/ui/dashboard/tests/modules/dashboard_test/src/Hook/ThemeHooks.php
@@ -5,11 +5,14 @@ declare(strict_types=1);
 namespace Drupal\farm_ui_dashboard_test\Hook;
 
 use Drupal\Core\Hook\Attribute\Hook;
+use Drupal\Core\StringTranslation\StringTranslationTrait;
 
 /**
  * Theme hook implementations for farm_ui_dashboard_test.
  */
 class ThemeHooks {
+
+  use StringTranslationTrait;
 
   /**
    * Implements hook_farm_dashboard_panes().
@@ -18,6 +21,7 @@ class ThemeHooks {
   public function farmDashboardPanes() {
     return [
       'dashboard_block' => [
+        'title' => $this->t('Test block title'),
         'block' => 'dashboard_test_block',
       ],
       'dashboard_view' => [